test(GameController): 게임 실행 기능 테스트 구현
- 기존 gameStart() 기능 추상화에 따른 Controller.test.js 파일로 테스트 구현체 이동
- 매 시행마다 각 자동차들의 상태 출력 테스트 케이스 추가
